pyproject.toml
[tool.poetry]
name = "weatherapp"
version = "0.1.0"
description = ""
authors = ["Nikita Savin <klaster26@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"
flask = "^3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
